system/vhba-module: Added (Virtual SCSI Host Bus adapter)
Re-added from 13.0 repository with a patch to bring everything
up to what's in the upstream svn repository (so it will build
on 2.6.33 and 2.6.34 kernels).

If the previous maintainer is no longer interested, we can look
at removing it for 13.2/14.0, but for now, let's eliminate the
missing dependency on it in our repo.

Signed-off-by: Robby Workman <rworkman@slackbuilds.org>