network/network-manager-applet: Included patch to fix fallback icon
If the applet was run in a window manager (e.g. fluxbox) that did
not set a gtk icon theme, then the system-wide fallback theme was
used, and the hicolor (fallback) theme does not contain the
gtk-dialog-error icons, so nm-applet refused to start.  This new
patch should fix that.  Thanks to Angel Montanez for reporting
this bug, and more importantly, thanks to NM's upstream developer,
Dan Williams, for addressing it so quickly.

Signed-off-by: Robby Workman <rworkman@slackbuilds.org>