graphics/mtpaint: Miscellaneous cleanups and a bugfix
Fix indentation. Move the sed before configuring, just as if it were a
patch. Remove useless strip and compress code and comment why. Make the
handbook handling better via an 'else' but also fix a bug where we were
in the wrong directory and copied the wrong COPYING/README files and
missed NEWS altogether. Now cd in a subshell and grab that COPYING as
COPYING.handbook. Bump BUILD because of this.

Signed-off-by: Robby Workman <rworkman@slackbuilds.org>