misc/dwdiff: Miscellaneous cleanups
Note in the README that the optional icu4c lib is available from SBo
(which it might not be, being optional). Move the NLS setting to the
rest of the settings so people who hand-edit don't have to look for
magic settings. Remove useless braces. Tweak strip boilerplate. Simplify
man page handling.

Signed-off-by: Robby Workman <rworkman@slackbuilds.org>