system/rox-filer: Miscellaneous cleanups
Clarify a point in the README that some users are interested in. Update
the doinst.sh to the new template. Remove unneeded comments. Comment why
there's no strip code. Fix indentation. Remove bashism (brace
expansion). Replace a 'cd...cd -' with a subshell. Tighten up other
one-line subshells so it's consistent with the rest of the one-line
subshells.

Signed-off-by: Robby Workman <rworkman@slackbuilds.org>