network/exim-html: Removed (hopefully temporarily)
In 4.73 the Exim developers changed the build process for the
documentation, which turned "exim-html" from a convenient standalone
package to just a dump of the exim.org Website with stylesheets and
images missing. So here we are...

- 4.72 is not up-to-date any more
- 4.73 and up is not convenient any more
- trying to fix it doesn't feel like "the Slackware way"
- full documentation is still there (spec.txt in the Exim package or
  online on exim.org and its mirrors), even without exim-html.

In conclusion I'd rather see exim-html removed from SBo for now.

The Exim git repository said it's not a permanent change, so if there
is a standalone doc again in a future version, I'll prepare a new
package.

Signed-off-by: Robby Workman <rworkman@slackbuilds.org>