libraries/libxml++: Noted mm-common in the README
This is an oversight on our part.  We (well, I) updated gtkmm
in our tree *after* all of the build testing was done, and the
previous gtkmm dependency chain didn't need mm-common, as the
relevant bits were shipped with glibmm.  However, the newer
glibmm no longer ships those, so we need mm-common now.
This may very well be a problem with some other c++ bindings
in our repo, so if you see any in the wild, shoot them and
send us the carcass.

Thanks to Luis Henrique F. Cardoso de Mello for the report.

Signed-off-by: Robby Workman <rworkman@slackbuilds.org>