desktop/dmenu: Revert to the dmenu_path script from 4.1.1
The dmenu_path in dmenu 4.3.1 seems to be picking up
non-executables on my system.  I was thinking of waiting
until upstream fixes this, but there is still some churn
going on regarding the dmenu_run/path scripts.  I figured
it would be better to get this small fix out there in the
mean time.

Signed-off-by: Robby Workman <rworkman@slackbuilds.org>