office/texlive: Included some patches from branch2011
Thanks to Karl Berry and the other TeXLive contributors for their
willingness to actively maintain a -stable branch of TeXLive with
only critical patches - this makes things much easier for distro
packagers and such.

This commit also includes some miscellaneous script fixes wrt
configure flags; thanks to Martin Ivanov for the tips.

Thanks also to Daniil Bratashov for making me aware of the build
problems that were fixed by the branch2011 patches (and to be
fair, he also attached split versions of those patches).

Finally, thanks also to Eduardo Sánchez for reporting the build
problems and testing the patches before I had a chance to do so.
Without feedback like this from our users, I'd be up a creek  :)

Signed-off-by: Robby Workman <rworkman@slackbuilds.org>