multimedia/kino: Removed (FTBFS and lack of time)
From: Pierre Cazenave <pwcazenave@gmail.com>
To: "SlackBuilds.org Users List" <slackbuilds-users@slackbuilds.org>
Subject: [Slackbuilds-users] Kino
Date: Mon, 09 Jan 2012 10:16:42 +0000

Kino has not been updated by upstream in some time, and getting it to
play nicely with the newest ffmpeg version is something for which I
don't have time. As it stands at the moment, it will not build with the
ffmpeg in the SBo repository. If anyone would like to take it over (and
its dependencies which I also maintain: libavc1394 and libiec61883),
they are welcome to do so. Otherwise, I suggest it is removed from the
repository as it doesn't work.

Signed-off-by: Robby Workman <rworkman@slackbuilds.org>