system/kvm-kmod: Fix build with non-running kernel and ...
Remove the depmod.d config file, as it doesn't appear to be needed
any more.  I'm leaving the bit of code that created it (it's only
commented out for now) since I don't know for *sure* that it's not
needed any more.  What I do know is that IF it IS still needed, the
previous file wasn't working anyway since the modules were installed
to a new location by upstream, and we (both maintainer and I) missed
that fact when updating the build script.  Oops :)

Signed-off-by: Robby Workman <rworkman@slackbuilds.org>