network/mumble: various build fixes and enhancements
- Fixed $SLKCFLAGS not being directly passed to qmake
- Included 'mumble-overly' script; this allows for the use of the
  mumble overlay
- Added COMPAT32 option for libmumble (preloaded overlay library)

The COMPAT32 option does build a 32-bit libmumble via the -m32
CFLAG, and I'm fairly certain this is taboo for SBo.  Obviously,
-m32 requires a multilibbed system, and not all x86_64 systems
have this setup. On x86_64 systems, mumble-overly will look for
a 32-bit libmumble to use, and it helps keep compatibility. Many
of the people using this are somewhat of "gamers", and use it on
games that are not native to the system (non-free and 32-bit only,
wine, etc).

Signed-off-by: Robby Workman <rworkman@slackbuilds.org>