libraries/simplegeneric: Require distribute instead of pysetuptools
Yes, pytsetuptools and distribute are mostly interchangeable (or at
least, distribute can supposedly *always* be used instead), but we
have seen a few cases where some things *require* distribute, so
there's a bit of a conflict when other things needed in the same
dep chain require pysetuputils.  It seems best to make everything
require distribute, and then we eliminate the potential conflicts.

Signed-off-by: Robby Workman <rworkman@slackbuilds.org>