development/jsctags: Removed (non-functional)
From: Audrius Kažukauskas <audrius@neutrino.lt>
To: slackbuilds-users@slackbuilds.org
Date: Tue, 11 Sep 2012 12:19:01 +0300

I only see development/jsctags on that list from the ones I maintain, so
I'm going to ask on this thread to simply drop it.  Yes, it does build
using the version of node from SBo, but it doesn't work afterwards.
Last time I tried to update it to the latest commit from GitHub repo, it
was a pain.  Doesn't look like Mozilla cares too much to make jsctags
easy for the packagers.

I'll look into it sometime in the future.  For now please remove it.
Thanks.

Signed-off-by: Robby Workman <rworkman@slackbuilds.org>