games/pengupop: Use Ubuntu's source archive.
Rather than modifying it and self-hosting, I'd prefer to use
the Ubuntu source archive.  On the off chance it goes missing,
or the Ubuntu folks express disagreement with this change, it
is also available on harrier in the "misc" directory of ftp.

This commit also silences some automake "missing" warnings.

Signed-off-by: Robby Workman <rworkman@slackbuilds.org>