network/etherpad-lite: Don't clobber system configs/logs on upgrade
This commit creates if needed (otherwise, updates timestamps on)
the log files during package install as opposed to creating them
inside the package itself - if they're present inside the package,
then a package upgrade clobbers the files on the installed system
if they happen to already exist.

This commit also installs the init script and a couple of config
files with .new extensions and then does the config() routine on
them (or preserve_perms() in the case of the init script).

The downside is that this will cause any .new'd files *and* the
logs to be removed on this upgrade; I hope this doesn't cause
much pain for users, but it's got to happen sooner or later,
so let's get it over with...

Signed-off-by: Robby Workman <rworkman@slackbuilds.org>