haskell/haskell-platform: Removed (build failure)
From: Niels Horn <niels.horn@slackbuilds.org>
To: c.willing@uq.edu.au,  SBo Admin List <slackbuilds-devel@slackbuilds.org>
Cc: vmj@linuxbox.fi
Subject: Re: [Slackbuilds-devel] New SBo Upload: haskell-platform
Date: Sat, 13 Apr 2013 22:45:27 -0300

On 03/03/2013 08:19 PM, SBo Admin List wrote:
> New upload: haskell-platform.tar.gz (c.willing@uq.edu.au)
> Maintainer: Christoph Willing
> Email: c.willing@uq.edu.au
> Category: Haskell
> Version: 2012.2.0.0 (db: 2012.2.0.0)
> Home Page: http://hackage.haskell.org/platform/
> Sources (32bit): http://lambda.haskell.org/platform/download/2012.4.0.0/haskell-platform-2012.4.0.0.tar.gz
> MD5SUMS (32bit): f1b1e6c94dddcdfbb36accafe74bdc0f
> Sources (64bit):
> MD5SUMS (64bit):
> Requires: hscolour
> Slackware: 14.0
> Submission md5: c2c1cf114dc42f1fd7e905f025de3085
> Comments: Update to 2012.4.0.0

Hi,

This has been sitting for over a month now in our pending queue, because
we found a problem and it took a while to investigate it better.

haskell-platform simply won't build with the ghc we have in our repo
(7.6.x). I tried to create a patch to get it to build, but had no luck.
I actually got quite far, but I'm not a very experienced haskell
programmer, so I had to give up.

After some discussion with the other admins, we decided that:
- we will not go back to ghc 7.4.x
- this submission will have to be rejected
- for now we will probably have to live without haskell-platform in our
repo, as the version we have there (2012.2.0.0) also does not build with
ghc 7.6.x

There seems to be a newer version of haskell-platform on the horizon
(2013.x) that hopefully solves this incompatibility.
Please resubmit when it comes out if it works with the newer ghc.

I'm cc'ing the maintainer of ghc. If you guys can come up with a
solution, please share it with us!

Signed-off-by: Robby Workman <rworkman@slackbuilds.org>