libraries/gstreamer1: Include orc as a dependency
It's not strictly required, but gstreamer and *all* of the
plugin packs strongly recommend it, as the codepaths used
without it will be slower.  User experience should be more
pleasing with this small library, so we'll hard-require it.
Of course, a user is still welcome to build without it if
they notice the cheat :-)

Signed-off-by: Robby Workman <rworkman@slackbuilds.org>