business/gnucash: Removed (build failure and etcetera)
This needs an older version of goffice than we have in
our repo, but at least a couple of other things in the
repo require the newer goffice.  Since we have to make
a choice, and since gnucash also insists on having all
of the deprecated gnome2 libraries, and because it's a
general pain anyway, the choice is an easy one to make
for us.  Isn't block text just wonderful looking here?

Signed-off-by: RobbyWorkman <rworkman@slackbuilds.org>