multimedia/mplayer-codecs64: Fixed VERSION string
Basically, the "essential" tarball contained dupes of what is
already in the other (earlier-dated) archive, so it is a useless
download that merely confuses the VERSION.  This appears to be
a downgrade, but in reality it's just a correction to what was
already being packaged.

Thanks to elseifthen <elseifthen@gmx.com> for the note with a
good explanation - much appreciated.

Signed-off-by: Robby Workman <rworkman@slackbuilds.org>