network/openconnect: Fix vpnc-script's test for /sbin/netconfig
Date: Mon, 8 Sep 2014 23:14:19 +0100
From: higuita <higuita@GMX.net>
To: rworkman@slackbuilds.org
Subject: openconnect slackbuild

I used your openconnect slackbuild to install it on my machine and
found a annoying problem.

The vpnc-script installed by openconnect script tried to
reconfigure slackware network configs files (rc.inet1.conf, hosts,
resolv.conf and others) because it is testing the existence of
/sbin/netconfig and use it... the problem is that is trying to
use it as the SuSE netconfig and is totally weird to get the
slackware netconfig vpn connection. Any user that don't cancel it
might mess up his network config without knowning

The solution is the attached patch, that bypass that test
on slackware machines. I will submit this also to the openconnect
so they can merge it too... but until that, please change the
slackbuild to apply this. If you like, i can send a patch for the
slackbuild too

Also, version on the slackbuild is 5.01 and openconnect
is already in version 6.00 and it works fine here... so no problem
in update it also.

--

Signed-off-by: Robby Workman <rworkman@slackbuilds.org>