libraries/libvirt: Updated for version 1.2.11.
This commit also modifies start() rc.libvirtd to be more
consistent with how Slackware's stock init scripts start
daemons, and a 2>/dev/null was added to the modprobe -r
command in stop() to hide useless errors when e.g. tun
module can't be removed because it's (still) in use (by
other processes).

Signed-off-by: Robby Workman <rworkman@slackbuilds.org>