network/kde-networkmanagement-openconnect: Updated for version 0.9.0.12.
There's a possibility that this won't work with the 0.9.0.9 version
of the parent package in Slackware 14.1, in which case we'll have to
revert openconnect back to the previous version. That's not really a
desirable move either though, since the newer version has support for
some new stuff that users want.  Let me know if there's a problem
with this and we'll go from there.

Signed-off-by: Robby Workman <rworkman@slackbuilds.org>