network/NetworkManager-openconnect: Updated for version 0.9.8.6.
As with the kde-networkmanagement-openconnect build, there's a chance
that this may not work due to it being a different version than the
NetworkManager and nm-applet version in Slackware 14.1, but this is
necessary due to the openconnect upgrade. If either this or the kde
applet don't work correctly, then let me know and we'll revert all
three of the upgrades.

Signed-off-by: Robby Workman <rworkman@slackbuilds.org>