Public www update: Wed Jan 13 03:59:58 UTC 2016
This will almost surely be the final regularly scheduled update
for the Slackware 14.1 branch here at SlackBuilds.org, since the
Slackware -current tree seems to be working its way toward a new
and shiny 14.2 release. We'll probably still cherry-pick a few
things (e.g. flash plugin) into the 14.1 branch prior to the 14.2
release, but understand that our primary focus from here on out
will be preparing everything for 14.2 so that we are ready when
it's released.

As always, thanks for your support over the years - it's a lot of
work, but I think I can speak for all of us admins when I say that
this is project is very rewarding and worth the time and effort.

Happy Slacking!

Signed-off-by: Robby Workman <rworkman@slackbuilds.org>