network/sslscan: Updated for version 1.10.2.
There are multiple forks of the original upstream (now dead and
fails to build). Arch uses this fork, but Debian and Fedora use
github.com/rbsec which is more active but wants a fresh git clone
of openssl for every build, so anyone who prefers rbsec is welcome
to fix that.

Actually, just use network/ssllabs-scan instead of this.

Signed-off-by: David Spencer <baildon.research@googlemail.com>