academic/sage: Downgraded to 7.2.
Built in python was compiled against UCS-4.
Slackware's readline is not linked to ncurses, causing
problems similar to here: https://bugs.archlinux.org/task/39731.

We will try to bump sage in the next Slackware release
when everything is fixed, including Python that is
compiled against UCS-4.

Adding a note for now.

Thanks to Mehran Hosseini for the bug report.

Signed-off-by: Willy Sudiarto Raharjo <willysr@slackbuilds.org>