Improve configure flags
Need to --enable-profiling to make the profile and cProfile from the standard library. They are used to debug the code.
In python 2.7.x from the default installation it was not necessary and was compiled by default.
Maybe you could also --enable-optimizations, but I don't know if its the case for a default make.