network/nss-mdns: Look in /run for avahi daemon socket
This fixes nss-mdns functionality since avahi was upgraded and
changed to use /run for its socket.

This build could be fixed by using --localstatedir=/ instead,
but it looks wrong (even if it's correct in this case), so I
decided to go with the patch instead.

Thanks to Lioh Moeller for the bug report.

Signed-off-by: Robby Workman <rworkman@slackbuilds.org>