[academics/root6] Create new SlackBuild for ROOT 6
The ROOT project switched its build system completely to cmake
starting from version 6.14/00 in February 2018, and in order to have
ROOT 6 available on upcoming Slackware 15 I decided to write my own
SlackBuild script. I took the existing root.SlackBuild script and
rewrote it in a way it can build new ROOT versions with cmake.
I successfully built ROOT 6.24/00 and ran some demos, read my own
analysis ntuples, and checked if python3 can import ROOT. It seems
everything is fine from my point of view, but of course there should
be done more testing (or improvement) for the script in future. This
version follows the Slackware 14.2 root.SlackBuild and installs
most of the files in /opt/ROOT-$VERSION (to be able to install both
root6 together root 5.34 on the same host, but I didn't check yet
how it works --- it needs to be done). Besides that, SB creates
/etc/profile.d/root6.sh which sources the original bin/thisroot.sh,
and /usr/doc/root6-$VERSION; README and LICENSE are also copied to
/opt/ROOT-$VERSION to run .license or .credits in ROOT shell
(C++ interpreter).

This script can't work with Slackware 14.2 (gcc 5.3). If you want to,
you need to change -DCMAKE_CXX_STANDARD=17 to 14. This is quite
important, because you can't have two different C++ standards in your
project using ROOT and in ROOT itself (or you need to wrap your
ROOT-dependent part). Maybe it is wise to have an extra SB parameter
(set by default to C++14), but then somebody should check how it works
on Slackware 14.2 (I don't have a proper computer for such tests).