gis/libspatialite: Updated for version 5.0.1
updated to support proj version 8
note that rttopo support is explicitly disabled as there is no librttopo
package in SBo.
Since this update is only to support the new proj, I leave it as an
exercise to the maintainer to add this dependency to SBo if they wish to
add the functionality.

Signed-off-by: Willy Sudiarto Raharjo <willysr@slackbuilds.org>