various: Use zulu jdk builds for REQUIRES.
As noted in the FAQ, you can still use Oracle's JDK for
building or running, but we'll default to the Zulu builds
in the REQUIRES since they are easily downloadable.

Zulu is a good default for us at SBo since it is available
for both 32 and 64 bit, freely downloadable and compatible
and they provide builds for all the LTS branches.

We don't currently have builds for Adoptium (previously
AdoptOpenJDK, not part of the Eclipse Foundation), but if
someone wants to submit builds for them, we can include them
too. Note though that they do not provide 32 bit builds.

Software should run with either Zulu or Oracle's JDK, or
indeed any other builds of OpenJDK, so if I have broken your
build with this change, apologies. This should be rare and
we can change the REQUIRES back to jdkX in that case and note
it in the README, but I do not expect this should be the case.

Signed-off-by: Andrew Clemons <andrew.clemons@gmail.com>

Signed-off-by: Willy Sudiarto Raharjo <willysr@slackbuilds.org>