ham/klog: Fix DOWNLOAD.
There is both a tag and a branch with the name 1.8.7, so the other
URL is ambigious. github recently started returning an HTML page
in this case indicating the error.

I have updated the URL to use the link directly to the tag and
added handling for both file names which can result.

Signed-off-by: Andrew Clemons <andrew.clemons@gmail.com>