system/gnu-unifont: Updated for version 14.0.02.
Prefer BDF input file when converting to OTB format

This is because BDF is the preferred input file format for the
fonttosfnt tool which is used when converting fonts to OTB format.

This is to allow using the fonts in Pango-based applications.
Support for bitmap fonts was dropped in Pango 1.44.

Signed-off-by: Willy Sudiarto Raharjo <willysr@slackbuilds.org>