libraries/adns: Fix empty package.
d189b4df7a broke this package such that we ended up with no
libs being installed. It seems the install-strip target does not
work properly. We'll simply use the template code to strip the
binaries instead.

I added back the deleting of the .a files which was removed in that
commit too, since they are back now that the package is not empty.

Signed-off-by: Andrew Clemons <andrew.clemons@gmail.com>

Signed-off-by: Willy Sudiarto Raharjo <willysr@slackbuilds.org>