system/gnu-unifont: Updated for version 15.0.01.
TTF format was removed from the default formats list as the fonts in
this format are not going to be supplied anymore by the upstream
starting from Unifont 16.0.01.

However, there still exist packages that depend on the TTF format of
this font.  Because of this, the removal of TTF format from the default
formats list has been reverted in this commit.

Signed-off-by: Willy Sudiarto Raharjo <willysr@slackbuilds.org>