python/python3-automat: Fix offline build.
If fixed this in 748383dcd0 by adding python-m2r as a dependency.
This is correct since upstream's `setup_requires` declares m2r as a
dependency:

https://github.com/glyph/automat/blob/v20.2.0/setup.py#L27

If m2r is not installed, the build will try to download it. If this
fails, the build fails. This means the package cannot be built
without internet at the moment.

```
distutils.errors.DistutilsError: Command '['/usr/bin/python3', '-m',
'pip', '--disable-pip-version-check', 'wheel', '--no-deps', '-w',
'/tmp/SBo/slackrepo.3Z4xrt/build_python3-automat/tmp0o3ejajj',
'--quiet', 'm2r']' returned non-zero exit status 1.
```

In 96170348cc, a change was made to make m2r optional, but this does
not work since it is still declared in setup.py. It should be
optional, since it is only used for formatting the contents of
README.md for use as long_description in setup and that is wrapped
in a try / except and will use the raw, unformatted text of
README.md if it is not found.

So for this change, I have patched out the declaration of m2r in
setup.py and you can now build the package without downloading
things as root during the build.

Signed-off-by: Willy Sudiarto Raharjo <willysr@slackbuilds.org>