python/python-debian: Fix gitlab tag handling.
The actual tag name is "debian/0.1.47". I've changed DOWNLOAD to
actually download the tag correctly while the previous URL resolved
to a commit. The two yield tarballs with different md5sums, but
the exact same contents.

The tag handling can be dropped again when this script is updated
the next time, since 0.1.48 and newer don't have a tag name prefixed
with ""debian".

Signed-off-by: Andrew Clemons <andrew.clemons@gmail.com>

Signed-off-by: Willy Sudiarto Raharjo <willysr@slackbuilds.org>