system/alacritty: Fix 64bit build.
I did not mean to stage this for release yet - I had made the
changes to the build, but had not yet built and tested it. I
inadvertently staged it to my branch and then pushed it out when I
was processing other github updates last week.

I have checked that this builds now and tested the resulting binary
here locally too.

Signed-off-by: Andrew Clemons <andrew.clemons@gmail.com>

Signed-off-by: Willy Sudiarto Raharjo <willysr@slackbuilds.org>