graphics/unpaper: Update for 7.0.0
Build system has moved to meson, so rewrite the SlackBuild based on the
standard meson template.

The manpage now requires "sphinx" to be built, which is a large python
package with many dependencies. Patch out the building of same by
default, and add the optional dep of Sphinx and the option SPHINX=true
for those that want it.

Update the README file to reflect this build option/optional dep, and
re-paragraph it to make it more readable.

Signed-off-by: Willy Sudiarto Raharjo <willysr@slackbuilds.org>