academic/root: Updated for version 6.32.04 (+fix 32 bit)
Added a patch from the 6.34 fixes branch to fix the issue #15738. It
shall be removed later when we switch to the 6.34 branch.

remove ld.so.conf and PATH manipulations

Wrapper scripts are created in /usr/bin for all ROOT binaries.
The trick does not work for python though, one needs to run
LD_LIBRARY_PATH=${ROOTSYS}/lib python3 in order to import ROOT
modules correctly.

Signed-off-by: Willy Sudiarto Raharjo <willysr@slackbuilds.org>